requirements.txt
# Core FastAPI dependencies
fastapi==0.115.4
starlette==0.41.2
uvicorn==0.32.1

# Database
sqlalchemy==2.0.40
sqlmodel==0.0.16
psycopg2-binary==2.9.10
pymongo==4.10.1

# Data validation and settings
pydantic==2.9.2
pydantic-settings==2.6.1

# Authentication & Security
PyJWT==2.9.0
passlib==1.7.4

# Configuration
python-decouple==3.8
python-dotenv==1.1.0

# HTTP requests
requests==2.32.3

# Date/Time utilities
pytz==2024.2

# Multi-part form data
python-multipart==0.0.18

# Type annotations (for older Python versions)
typing_extensions==4.13.2
